# docs/swagger.yaml
definitions:
...
  controllers.SellerDTO:
    properties:
      address:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        maxLength: 16
        type: string
    required:
    - address
    - name
    - phone
    - username
    type: object
  controllers.SellerUpdateDTO:
    properties:
      address:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - address
    - name
    - phone
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /Sellers/{id}:
    delete:
      description: Delete a Seller by id.
      parameters:
      - description: Seller id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete one Seller.
      tags:
      - Seller
    put:
      description: Update Seller by id.
      parameters:
      - description: seller id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an Seller
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.SellerUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update Seller.
      tags:
      - Seller
...
  /sellers/{id}:
    get:
      description: Get a Seller by id.
      parameters:
      - description: Seller id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Seller.
      tags:
      - Seller
  /sellers/{id}/products:
    get:
      description: Get all Products by Seller Id.
      parameters:
      - description: Seller id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Products by seller id.
      tags:
      - Seller
  /sellers/register:
    post:
      description: Creating a new Seller.
      parameters:
      - description: the body to create a new Seller
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.SellerDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register New Seller.
      tags:
      - Seller
swagger: "2.0"